issue: 1284069 Modify map_device logic
These changes allow to do net_dev_val initialization
inside constructor basing on struct ifaddrs information.
So net_dev_val() ctor is able to check if current interface
can be used to offload traffic and prepare object for
usage in case success decision.
